refactor to eliminate duplicate split initialization and evaluation on the client
